requirements.txt
requests
beautifulsoup4
pandas
numpy
httplib2
google-api-python-client==1.7.2
google-auth==1.8.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
oauth2client
gspread
gspread-dataframe